# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Recruitment Registry",
        "license": {
            "name": "Apache 2.0"
        }
    },
    "basePath": "/api/v1.0",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/answers": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "description": "This persists answers to survey questions.  Old answers are soft deleted.  For each question for which there is no answer property, any existing answer will be soft deleted.",
                "summary": "Creates answers",
                "operationId": "createAnswers",
                "security": [{
                    "participant": []
                }],
                "parameters": [{
                    "name": "surveyAnswers",
                    "in": "body",
                    "description": "Answers to the survey",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/surveyAnswers"
                    }
                }, {
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "201": {
                        "$ref": "#/responses/emptySuccess"
                    },
                    "400": {
                        "$ref": "#/responses/badRequest"
                    },
                    "401": {
                        "$ref": "#/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/responses/forbidden"
                    },
                    "422": {
                        "$ref": "#/responses/unprocessable"
                    },
                    "500": {
                        "$ref": "#/responses/internal"
                    }
                }
            },
            "get": {
                "summary": "Gets answers",
                "operationId": "getAnswers",
                "security": [{
                    "participant": []
                }],
                "parameters": [{
                    "name": "surveyId",
                    "in": "query",
                    "description": "Survey id",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/linkedAnswer"
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/responses/badRequest"
                    },
                    "401": {
                        "$ref": "#/responses/unauthorized"
                    },
                    "403": {
                        "$ref": "#/responses/forbidden"
                    },
                    "422": {
                        "$ref": "#/responses/unprocessable"
                    },
                    "500": {
                        "$ref": "#/responses/internal"
                    }
                }
            }
        },
...
        "/consent-types": {
            "x-swagger-router-controller": "consent-type.controller",
            "post": {
                "summary": "Creates new consent type",
                "description": "Each consent type describes a consent section that users have to accept for various registry functionality.  There can be multiple consent sections with the same type but only one of them can be active at one time.",
                "operationId": "createConsentType",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "newConsentType",
                    "in": "body",
                    "description": "New consent type.",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "required": ["name", "title", "type"],
                        "properties": {
                            "name": {
                                "type": "string",
                                "minLength": 1
                            },
                            "title": {
                                "type": "string",
                                "minLength": 1
                            },
                            "type": {
                                "type": "string",
                                "minLength": 1
                            }
                        },
                        "additionalProperties": false
                    }
                }],
                "responses": {
                    "201": {
                        "$ref": "#/responses/idSuccess"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "get": {
                "summary": "Lists consent types",
                "operationId": "listConsentTypes",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/consentType"
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-types/{id}": {
            "x-swagger-router-controller": "consent-type.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets the consent type",
                "operationId": "getConsentType",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "$ref": "#/definitions/consentType"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "delete": {
                "summary": "Deletes the consent type",
                "operationId": "deleteConsentType",
                "security": [{
                    "admin": []
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-types/text/{language}": {
            "x-swagger-router-controller": "consent-type.controller",
            "patch": {
                "summary": "Updates or translates consent type title",
                "operationId": "updateConsentTypeText",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "path",
                    "description": "Language code",
                    "required": true,
                    "type": "string",
                    "minLength": 2
                }, {
                    "name": "textUpdate",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/consentTypeTextUpdate"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successul response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-documents": {
            "x-swagger-router-controller": "consent-document.controller",
            "post": {
                "summary": "Creates a consent section",
                "operationId": "createConsentDocument",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "newConsentDocument",
                    "in": "body",
                    "description": "Content text",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "required": ["typeId", "content"],
                        "properties": {
                            "typeId": {
                                "type": "integer"
                            },
                            "content": {
                                "type": "string",
                                "minLength": 1
                            },
                            "updateComment": {
                                "type": "string",
                                "minLength": 1
                            }
                        },
                        "additionalProperties": false
                    }
                }],
                "responses": {
                    "201": {
                        "$ref": "#/responses/idSuccess"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-documents/{id}": {
            "x-swagger-router-controller": "consent-document.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets consent section content",
                "operationId": "getConsentDocument",
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "object",
                            "required": ["id", "typeId", "content"],
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "typeId": {
                                    "type": "integer"
                                },
                                "content": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "updateComment": {
                                    "type": ["string", "null"]
                                }
                            },
                            "additionalProperties": false
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-documents/update-comments/{typeId}": {
            "x-swagger-router-controller": "consent-document.controller",
            "parameters": [{
                "name": "typeId",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets update comments for consent document",
                "operationId": "getUpdateCommentHistory",
                "parameters": [{
                    "name": "language",
                    "in": "query",
                    "description": "Language code",
                    "required": false,
                    "type": "string",
                    "minLength": 2
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": ["string", "null"]
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/consent-documents/text/{language}": {
            "x-swagger-router-controller": "consent-document.controller",
            "patch": {
                "summary": "Translates consent type document",
                "operationId": "updateConsentDocumentText",
                "security": [{
                    "admin": []
                }],
                "parameters": [{
                    "name": "language",
                    "in": "path",
                    "description": "Language code",
                    "required": true,
                    "type": "string",
                    "minLength": 2
                }, {
                    "name": "textUpdate",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/consentDocumentTextUpdate"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successul response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "admin": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        },
        "participant": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        },
        "self": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "responses": {
        "unauthorized": {
            "description": "User is not authenticated.",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
        "forbidden": {
            "description": "User does not have permission.",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
        "badRequest": {
            "description": "Invalid request parameters.",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
        "unprocessable": {
            "description": "Request cannot be fullfilled",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
        "genericError": {
            "description": "Error response",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
        "emptySuccess": {
            "description": "Successful response.",
            "schema": {
                "$ref": "#/definitions/emptyObject"
            }
        },
        "idSuccess" : {
            "description": "Successful response",
            "schema": {
                "description": "Id",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                }
            }
        },
        "internal": {
            "description": "Unexpected error.",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
...
    },
    "definitions": {
        "error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "answer": {
            "type": "object",
            "properties": {
                "choice": {
                    "type": "integer",
                    "minimum": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": ["id"],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "boolValue": {
                                "type": "boolean"
                            },
                            "textValue": {
                                "type": "string"
                            }
                        },
                        "maxProperties": 2
                    },
                    "minItems": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                }
            },
            "minProperties": 1,
            "maxProperties": 1
        },
...
        "linkedAnswer": {
            "type": "object",
            "required": [
                "questionId"
            ],
            "properties": {
                "questionId": {
                    "type": "integer",
                    "minimum": 1
                },
                "answer": {
                    "$ref": "#/definitions/answer"
                }
            }
        },
...
        "surveyAnswers": {
            "type": "object",
            "required": [
                "surveyId",
                "answers"
            ],
            "properties": {
                "surveyId": {
                    "type": "integer",
                    "minimum": 1
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/linkedAnswer"
                    },
                    "minItems": 1
                }
            }
        },
...
        "emptyObject": {
            "type": "object",
            "maxProperties": 0
        },
...
        "consentType": {
            "type": "object",
            "required": ["id", "name", "title", "type"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                },
                "type": {
                    "type": "string",
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "consentTypeTextUpdate": {
            "type": "object",
            "required": ["id", "title"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
        "consentDocumentTextUpdate": {
            "type": "object",
            "required": ["id", "content"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "content": {
                    "type": "string",
                    "minLength": 1
                },
                "updateComment": {
                    "type": ["string", "null"],
                    "minLength": 1
                }
            },
            "additionalProperties": false
        },
...
    }
}